+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @ai-sdk/vue
 
+## 3.0.0-beta.109
+
+### Patch Changes
+
+- Updated dependencies [79ba99f]
+  - ai@6.0.0-beta.109
+
+## 3.0.0-beta.108
+
+### Patch Changes
+
+- ai@6.0.0-beta.108
+
+## 3.0.0-beta.107
+
+### Patch Changes
+
+- Updated dependencies [8445d70]
+  - ai@6.0.0-beta.107
+
+## 3.0.0-beta.106
+
+### Patch Changes
+
+- ai@6.0.0-beta.106
+
 ## 3.0.0-beta.105
 
 ### Patch Changes